some security added

+++ controllers/userController.js
@@ -2,8 +2,10 @@ const { User } = require("../models/userModel");
 
 exports.fetchUser = async (req, res) => {
   try {
-    const user = await User.find({}).select("-password -salt").exec();
-    res.status(200).json(user);
+    if (req.user.role == "Admin") {
+      const user = await User.find({}).select("-password -salt").exec();
+      res.status(200).json(user);
+    }
   } catch (err) {
     console.log(err);
     res.status(400).json(err);
@@ -12,8 +14,10 @@ exports.fetchUser = async (req, res) => {
 
 exports.deleteUser = async (req, res) => {
   try {
-    const user = await User.findByIdAndDelete(req.params.id);
-    res.status(200).json(user);
+    if (req.user.role == "Admin") {
+      const user = await User.findByIdAndDelete(req.params.id);
+      res.status(200).json(user);
+    }
   } catch (err) {
     console.log(err);
     res.status(400).json(err);
@@ -23,14 +27,16 @@ exports.deleteUser = async (req, res) => {
 exports.updateUserById = async (req, res) => {
   const { id } = req.params;
   try {
-    const user = await User.findByIdAndUpdate(id, req.body, {
-      new: true,
-    });
-    const updateUser = await user.save();
-    const element = await User.findById(updateUser.id)
-      .select("-password -salt")
-      .exec();
-    res.status(200).json(element);
+    if (req.user.role === "Admin") {
+      const user = await User.findByIdAndUpdate(id, req.body, {
+        new: true,
+      });
+      const updateUser = await user.save();
+      const element = await User.findById(updateUser.id)
+        .select("-password -salt")
+        .exec();
+      res.status(200).json(element);
+    }
   } catch (err) {
     console.log(err);
     res.status(200).json(err);
